ParrotMainCamera: setZoom now takes a double zoom_level.
This is to harmonize with the fact that for thrift there is no float,
but only double. setZoom for thrift is not available yet but this
harmonization is necessary to achieve it.

+++ parrot-backend/android/app/src/main/java/com/drohub/ParrotHelpers/Peripherals/ParrotMainCamera.java
@@ -79,7 +79,7 @@ public class ParrotMainCamera implements IParrotPeripheral {
         TOO_FAST
     }
 
-    public ZoomResult setZoom(float zoom_level) {
+    public ZoomResult setZoom(double zoom_level) {
         if (!_priv.getZoom().isAvailable())
             return ZoomResult.BAD;
         if (System.currentTimeMillis() - last_zoom_set < 250)
